+++ test/endpoint.js
@@ -113,6 +113,10 @@ describe('stemmers', function() {
   });
 });
 
+describe('tag', function() {
+  require('./tag/averaged-perceptron.js');
+});
+
 describe('tokenizers', function() {
   describe('hyphenation', function() {
     require('./tokenizers/hyphenation/liang.js');